requirements.txt
alembic==1.8.1
aniso8601==9.0.1
anyio==3.6.1
attrs==22.1.0
certifi @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_05nm_gqf36/croots/recipe/certifi_1663615689491/work/certifi
click==8.1.3
fastapi==0.85.0
fastapi-utils==0.2.1
Flask==2.1.2
Flask-Migrate==3.1.0
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.3
idna==3.4
importlib-metadata==4.12.0
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.16.0
Mako==1.2.3
MarkupSafe==2.1.1
migrate==0.3.8
numpy==1.23.3
packaging==21.3
pandas==1.5.0
pluggy==1.0.0
py==1.11.0
pydantic==1.10.2
pyparsing==3.0.9
pyrsistent==0.18.1
pytest==7.1.3
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.2.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
starlette==0.20.4
tomli==2.0.1
typing_extensions==4.3.0
Werkzeug==2.1.2
zipp==3.8.1
